# Optimize trait loop evaluation