--- a/lib/services/traits.go
+++ b/lib/services/traits.go
@@ -104,7 +104,7 @@ func traitsToRoles(ms types.TraitMappingSet, traits map[string][]string, collect
 					switch {
 					case err != nil:
 						// this trait value clearly did not match, move on to another
-						mismatched = append(mismatched, traitValue)
+						mismatched = append(mismatched, fmt.Sprintf("%q", traitValue))
 						continue TraitLoop
 					case outRole == "":
 					case outRole != "":
@@ -124,9 +124,20 @@ func traitsToRoles(ms types.TraitMappingSet, traits map[string][]string, collect
 
 			// warn at most maxMismatchedTraitValuesWarned trait values to prevent huge warning lines
 			if len(mismatched) > maxMismatchedTraitValuesWarned {
-				warnings = append(warnings, fmt.Sprintf("%d trait values did not match expression %q: %s (first %d values)", len(mismatched), mapping.Value, mismatched[0:maxMismatchedTraitValuesWarned], maxMismatchedTraitValuesWarned))
+				warnings = append(warnings, fmt.Sprintf(
+					"%d trait value(s) did not match expression %q: [%s] (first %d values)",
+					len(mismatched),
+					mapping.Value,
+					mismatched[0:maxMismatchedTraitValuesWarned],
+					maxMismatchedTraitValuesWarned,
+				))
 			} else if len(mismatched) > 0 {
-				warnings = append(warnings, fmt.Sprintf("%d trait values did not match expression %q: %s", len(mismatched), mapping.Value, mismatched))
+				warnings = append(warnings, fmt.Sprintf(
+					"%d trait value(s) did not match expression %q: %s",
+					len(mismatched),
+					mapping.Value,
+					mismatched,
+				))
 			}
 		}
 	}