@codingllama 2022-12-06
> Might be nice to add a benchmark here as well, which would show how much of an improvement this change makes, as well as make it easier to detect performance regressions in the future.

+1 , let's get some numbers in. It's hard to assess how well this works just by reading it.