@codingllama 2022-12-07
> @zmb3 @codingllama [ce86dc0](https://github.com/gravitational/teleport/commit/ce86dc0777bdaaae4e8d28488e911f81173110de) adds a benchmark using the test cases we already had for `TestOIDCMapping` (this made the PR harder to review though).
> 
> I ran the benchmark both with `go test` and `go test -v` to try to also account for the `log.Debug` calls. The PR description now contains some benchmark results.

Many thanks! LGTM.